Домашнее задание №5. Модуляция. Правки +
