+++ js/main.js
@@ -1,5 +1,3 @@
-import {OBJECT_COUNT} from './data.js';
-import {createObj} from './util.js';
+import {createSimilarObjects} from './create-similar-objects.js';
 
-const createSimilarObjects = () => new Array(OBJECT_COUNT).fill(null).map(() => createObj());
 createSimilarObjects();
